Changed visual stuff a bit, added new button on TransactionsView to show current info.

+++ GUI/CustomerWallet/AddWalletViewModel.cs
@@ -43,7 +43,7 @@ namespace GUI.CustomerWallet
 
             set
             {
-                if (wallet.StartBalance != (double)value)
+                if (wallet.StartBalance != value)
                 {
                     wallet.StartBalance = (double)value;
                     OnPropertyChanged();
